moved proto/sign to lib/sign
