+++ app/sign/sign.go
@@ -4,7 +4,7 @@ import (
 	"github.com/dedis/cothority/lib/app"
 	dbg "github.com/dedis/cothority/lib/debug_lvl"
 	"github.com/dedis/cothority/lib/graphs"
-	"github.com/dedis/cothority/proto/sign"
+	"github.com/dedis/cothority/lib/sign"
 	"io/ioutil"
 	"os"
 	"time"
