+++ lib/conode/callbackstamper.go
@@ -12,7 +12,7 @@ import (
 	"github.com/dedis/cothority/lib/coconet"
 	"github.com/dedis/cothority/lib/hashid"
 	"github.com/dedis/cothority/lib/proof"
-	"github.com/dedis/cothority/proto/sign"
+	"github.com/dedis/cothority/lib/sign"
 	"github.com/dedis/crypto/abstract"
 	"github.com/dedis/cothority/lib/cliutils"
 	"net"
