+++ lib/sign/basicSig.go
