+++ lib/sign/collectiveSigning.go
