+++ lib/sign/consensus.go
