+++ lib/sign/constants.go
