+++ lib/sign/errors.go
