+++ lib/sign/interfaces.go
