+++ lib/sign/peer.go
