+++ lib/sign/signingNode.go
