+++ lib/sign/snutils.go
