+++ lib/sign/testdata/extcpconf.json
